GSvar: fixed bug in variant quality coloring (floats above 20 were colored yellow because .toInt() was used to convert the numeric string to a number)

+++ src/GSvar/VariantDetailsDockWidget.cpp
@@ -351,7 +351,7 @@ void VariantDetailsDockWidget::setAnnotation(QLabel* label, const VariantList& v
 
 				if (part.startsWith("QUAL="))
 				{
-					if (part.mid(5).toInt()<20)
+					if (part.mid(5).toDouble()<20)
 					{
 						part = formatText(part, YELLOW);
 					}
